Update to regex-automata v0.4 (#5)
This updates to use regex-automata v0.4, bumping the crate to 0.2 since
regex-automata is a public dependency of the crate. All tests in this
crate continue to pass.

This removes the `ToMatcher` trait in favor of directly using the
upstream Automaton trait, it didn't seem like that trait was really
being used. As of the current PR, there's no way to construct Patterns
with the sparse DFA (this seems to have been true of the prior version
of the crate too), but all the functionality is generic over
[regex_automata::Automaton](https://docs.rs/regex-automata/latest/regex_automata/dfa/trait.Automaton.html),
so just adding constructs for the lazy DFA would be enough to make it
possible to use.

This also reduces the enabled features on regex-automata to not include
unicode features. It's not obvious that downstream crates (in particular
tracing-subscriber) want to pay the binary size cost for having those
enabled in env-filter regexes, and this leaves the choice up to them.

Closes #3